+++ urbanairship-react-native/ios/UARCTModule/UARCTModuleVersion.m
@@ -4,7 +4,7 @@
 
 @implementation UARCTModuleVersion
 
-NSString *const airshipModuleVersionString = @"13.2.1";
+NSString *const airshipModuleVersionString = @"14.0.0";
 
 + (nonnull NSString *)get {
     return airshipModuleVersionString;