+++ urbanairship-react-native/js/SubscriptionLists.ts
@@ -2,8 +2,11 @@
 
 'use strict';
 
+import { SubscriptionScope } from "./UrbanAirship";
+
 export interface SubscriptionLists {
     channel?: string[]
+    contact?: Map<string, SubscriptionScope[]>
 }
 
-export type SubscriptionListType = "channel"
+export type SubscriptionListType = "channel"|"contact"